Refactoring unit tests 1/n
